requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
tavily-python>=0.3.0
python-dotenv>=1.0.0
